+++ packages/main/internals/History.js
@@ -1,32 +1,24 @@
-import _reduce from 'ramda/src/reduce'
+import reduce from 'ramda/src/reduce'
 import delay from './delay'
-// import map from 'ramda/src/map'
 
-const history = (state, reducer, _rerender) => {
-  const reduce = _reduce(reducer)
-  // this.state = state
-  const subscriptions = []
-  let delta = []
-  const rerender = () => _rerender(state)
+export default class History {
+  constructor (state, reducer, rerender) {
+    this.reduce = reduce(reducer)
+    this.state = state
+    this.delta = []
+    this.rerender = () => rerender(this.state)
+  }
 
-  const subscribe = (subscription) =>
-    subscriptions.push(subscription)
+  concat () {
+    this.state = this.reduce(this.state, this.delta)
+    const dom = this.rerender()
+    this.delta = []
 
-  const concat = () => {
-    state = reduce(state, delta)
-    const dom = rerender()
-    delta = []
-
-    // map((subscription) => subscription(this), this.subscriptions)
     return dom
   }
 
-  const push = (action) => {
-    delta.push(action)
-    return delay(() => concat())
+  push (action) {
+    this.delta.push(action)
+    return delay(() => this.concat())
   }
-
-  return { subscribe, push }
 }
-
-export default history