+++ packages/main/test/index.js
@@ -182,7 +182,7 @@ test('main: calls storage functions when value changes', (assert) => {
 })
 
 test('main: history.undo/redo', (assert) => {
-  const { history } = main(identity, { model: 'Heraclitus', historyFactory: TraversableHistory })
+  const { history } = main(identity, { model: 'Heraclitus', historyClass: TraversableHistory })
 
   history.push({ type: 'name', payload: 'Parmenides' })
   const dom = history.push({ type: 'name', payload: 'Anaximander' })
@@ -199,7 +199,7 @@ test('main: history.undo/redo', (assert) => {
 })
 
 test('main: history.go', (assert) => {
-  const { history } = main(identity, { model: 'Anaximenes', historyFactory: TraversableHistory })
+  const { history } = main(identity, { model: 'Anaximenes', historyClass: TraversableHistory })
 
   history.push({ type: 'name', payload: 'Plutarchus' })
   history.push({ type: 'name', payload: 'Plotinus' })
@@ -219,7 +219,7 @@ test('main: mouse events', (assert) => {
   forEach((mouseEvent) => {
     const view = ['button', {[mouseEvent]: {add: 1}}]
 
-    const { dom, history } = main(view, { historyFactory: TraversableHistory })
+    const { dom, history } = main(view, { historyClass: TraversableHistory })
     mouseEvents[mouseEvent](dom)
     assert.deepEqual(
       history.current,
@@ -235,7 +235,7 @@ test('main: keyboard events', (assert) => {
   forEach((keyboardEvent) => {
     const view = ['button', {[keyboardEvent]: {get: valueAndKeyCode}}]
     const randomKeyCode = parseInt(Math.random() * 20 + 1)
-    const { dom, history } = main(view, { historyFactory: TraversableHistory })
+    const { dom, history } = main(view, { historyClass: TraversableHistory })
     keyboardEvents[keyboardEvent](dom, randomKeyCode, keyboardEvent)
     assert.deepEqual(
       history.current,
@@ -251,7 +251,7 @@ test('main: form events', (assert) => {
   forEach((formEvent) => {
     const view = ['form', {[formEvent]: {data: naiveSerialize}}]
 
-    const { dom, history } = main(view, { historyFactory: TraversableHistory })
+    const { dom, history } = main(view, { historyClass: TraversableHistory })
     formEvents[formEvent](dom, {name: 'Freud', age: 33})
     assert.deepEqual(
       history.current,
@@ -266,7 +266,7 @@ test('main: form events', (assert) => {
 test('main: input events', (assert) => {
   forEach((inputEvent) => {
     const view = ['input', {[inputEvent]: {value: value}}]
-    const { dom, history } = main(view, { historyFactory: TraversableHistory })
+    const { dom, history } = main(view, { historyClass: TraversableHistory })
     inputEvents[inputEvent](dom, inputEvent)
     assert.deepEqual(
       history.current,